+++ .env.sample
@@ -1,3 +1,3 @@
 ENVIRONMENT=development
-DATABASE_NAME=my-tutorial-db
-DATABASE_COLLECTION_NAME=my-collection
+DATABASE_NAME=
+DATABASE_COLLECTION_NAME=
