+++ .eslintrc
@@ -2,17 +2,17 @@
   "plugins": [
     "prettier"
   ],
-    "root": true,
-      "extends": [
-        "eslint:recommended",
-        "airbnb-base",
-        "prettier"
-      ],
-        "env": {
+  "root": true,
+  "extends": [
+    "eslint:recommended",
+    "airbnb-base",
+    "prettier"
+  ],
+  "env": {
     "commonjs": true,
-      "es2020": true,
-        "node": true,
-          "jest": true
+    "es2020": true,
+    "node": true,
+    "jest": true
   },
   "parserOptions": {
     "ecmaVersion": 11
@@ -22,18 +22,20 @@
     ".devcontainer",
     "dist"
   ],
-    "rules": {
+  "rules": {
     "semi": [
       "warn",
       "always"
     ],
-      "quotes": [
-        "error",
-        "single"
-      ],
-        "no-console": [
-          "warn"
-        ],
-          "no-return-await": ["warn"]
+    "quotes": [
+      "error",
+      "single"
+    ],
+    "no-console": [
+      0
+    ],
+    "no-return-await": [
+      "warn"
+    ]
   }
 }
