+++ src/data.js
@@ -1,81 +1,64 @@
 const { MongoClient } = require('mongodb');
 const ObjectId = require('mongodb').ObjectID;
+// read .env file
+require('dotenv').config();
+
+/* eslint no-return-await: 0 */
+
+const DATABASE_URL = process.env.DATABASE_URL
+    ? process.env.DATABASE_URL
+    : 'mongodb://localhost:27017';
+const DATABASE_NAME = process.env.DATABASE_NAME || 'my-tutorial-db';
+const DATABASE_COLLECTION_NAME =
+    process.env.DATABASE_COLLECTION_NAME || 'my-collection';
+
+let mongoConnection = null;
+let db = null;
+
+/* eslint no-console: 0 */
+console.log(`DB:${DATABASE_URL}`);
 
 const insertDocuments = async (
-    db,
-    collectionName = 'documents',
     documents = [{ a: 1 }, { a: 2 }, { a: 3 }]
 ) => {
-    
     // check params
-    if (!db || !collectionName || !documents)
+    if (!db || !documents)
         throw Error('insertDocuments::missing required params');
 
     // Get the collection
-    const collection = await db.collection(collectionName);
-    
+    const collection = await db.collection(DATABASE_COLLECTION_NAME);
+
     // Insert some documents
-    await collection.insertMany(documents);
+    return await collection.insertMany(documents);
 };
 const findDocuments = async (
-    db,
-    collectionName = 'documents',
     query = { a: 3 }
 ) => {
     
     // check params
-    if (!db || !collectionName)
+    if (!db)
         throw Error('findDocuments::missing required params');
 
     // Get the collection
-    const collection = await db.collection(collectionName);
+    const collection = await db.collection(DATABASE_COLLECTION_NAME );
 
     // find documents
-    await collection.find(query).toArray();
+    return await collection.find(query).toArray();
 };
 
 const removeDocuments = async (
-    db,
-    collectionName = 'documents',
     docFilter = {}
 ) => {
     
     // check params
-    if (!db || !collectionName)
+    if (!db )
         throw Error('removeDocuments::missing required params');
 
     // Get the documents collection
-    const collection = await db.collection(collectionName);
+    const collection = await db.collection(DATABASE_COLLECTION_NAME);
 
     // Delete document
-    await collection.deleteMany(docFilter);
-};
-
-const removeDocument = async (db, collectionName = 'documents', id) => {
-    
-    // check params
-    if (!db || !collectionName || !id)
-        throw Error('removeDocument::missing required params');
-
-    // Get the documents collection
-    const collection = await db.collection(collectionName);
-
-    // Delete document
-    await collection.deleteOne({ _id: ObjectId(id) });
-};
-
-const indexCollection = async (
-    db,
-    collectionName = 'documents',
-    index = { a: 1 }
-) => {
-    
-    // check params
-    if (!db || !collectionName || !index)
-        throw Error('indexCollection::missing required params');
-
-    // create index
-    await db.collection(collectionName).createIndex(index, null);
+    return await collection.deleteMany(docFilter);
 };
 
 const connect = async (url) => {
@@ -85,20 +68,33 @@ const connect = async (url) => {
 
     return MongoClient.connect(url, { useUnifiedTopology: true });
 };
-const disconnect = async (client) => {
-    
-    // check params
-    if (!client) throw Error('disconnect::missing required params');
+/* 
+eslint consistent-return: [0, { "treatUndefinedAsUnspecified": false }]
+*/
+const connectToDatabase = async () => {
+    try {
+        if (!DATABASE_URL || !DATABASE_NAME) {
+            console.log('DB required params are missing');
+            console.log(`DB required params DATABASE_URL = ${DATABASE_URL}`);
+            console.log(`DB required params DATABASE_NAME = ${DATABASE_NAME}`);
+        }
 
-    client.close();
+        mongoConnection = await connect(DATABASE_URL);
+        db = mongoConnection.db(DATABASE_NAME);
+
+        console.log(`DB connected = ${!!db}`);
+        
+        return !!db;
+
+    } catch (err) {
+        console.log('DB not connected - err');
+        console.log(err);
+    }
 };
-
 module.exports = {
     insertDocuments,
     findDocuments,
-    removeDocument,
     removeDocuments,
-    indexCollection,
-    connect,
-    disconnect,
+    ObjectId,
+    connectToDatabase
 };
