+++ test/data-integration.test.js
@@ -1,28 +1,16 @@
 const data = require('../src/data');
 require('dotenv').config();
 
-// Test requires connection to DB
-const DATABASE_URL = process.env.DATABASE_URL
-    ? process.env.DATABASE_URL
-    : 'mongodb://localhost:27017';
-const DATABASE_NAME = process.env.DATABASE_NAME || 'my-tutorial-db';
-const DATABASE_COLLECTION_NAME =
-    process.env.DATABASE_COLLECTION_NAME || 'my-collection';
-
 const DATABASE_DOCS = [{ a: 1 }, { a: 2 }, { a: 3 }];
 const ALL_DOCS = {};
 
 describe('mongoDB native API', () => {
     test('integration with DB', async (done) => {
-        const dbProcess = async (url, dbName, collectionName, docs) => {
-            const client = await data.connect(url);
+        const dbProcess = async (docs) => {
+            await data.connectToDatabase();
 
-            const db = await client.db(dbName);
-            expect(db).not.toBe(undefined);
-
+            // insert new docs
             const insertResult = await data.insertDocuments(
-                db,
-                collectionName,
                 docs
             );
             expect(insertResult).not.toBe(undefined);
@@ -31,19 +19,15 @@ describe('mongoDB native API', () => {
             const id = insertResult.insertedIds[0].toString();
 
             const findResult = await data.findDocuments(
-                db,
-                collectionName,
                 ALL_DOCS
             );
             expect(findResult.length).not.toBe(0);
 
             const count = findResult.length;
 
-            await data.removeDocument(db, collectionName, id);
+            await data.removeDocuments({ _id: data.ObjectId(id) });
 
             const findResult2 = await data.findDocuments(
-                db,
-                collectionName,
                 ALL_DOCS
             );
             expect(findResult2.length).not.toBe(0);
@@ -51,26 +35,17 @@ describe('mongoDB native API', () => {
             expect(findResult2.length).toEqual(count - 1);
 
             const removeResult = await data.removeDocuments(
-                db,
-                collectionName,
                 ALL_DOCS
             );
             expect(removeResult).not.toBe(undefined);
 
             const findResult3 = await data.findDocuments(
-                db,
-                collectionName,
                 ALL_DOCS
             );
             expect(findResult3.length).toBe(0);
-
-            await data.disconnect(client);
         };
 
         dbProcess(
-            DATABASE_URL,
-            DATABASE_NAME,
-            DATABASE_COLLECTION_NAME,
             DATABASE_DOCS
         )
             .then(() => done())
